use f32

+++ src/logistic_map.rs
@@ -1,14 +1,14 @@
 
 
-const R: f64 = 4.0;
+const R: f32 = 4.0;
 
 struct LogisticMapOptions {
-    x: f64,
-    r: f64,
+    x: f32,
+    r: f32,
     size: usize,
 }
 
-fn generate_map(opt: LogisticMapOptions) -> Vec<f64> {
+fn generate_map(opt: LogisticMapOptions) -> Vec<f32> {
     let mut pixels = Vec::with_capacity(opt.size);
     let mut x = opt.x;
 
@@ -27,7 +27,7 @@ fn argsort<T: Ord>(vec: &Vec<T>) -> Vec<usize> {
     indices
 }
 
-fn confuse_xor(im: &Vec<[u8; 3]>, enc_map: &Vec<f64>) -> Vec<[u8; 3]> {
+fn confuse_xor(im: &Vec<[u8; 3]>, enc_map: &Vec<f32>) -> Vec<[u8; 3]> {
     let mut enc_pixels = Vec::with_capacity(im.len());
 
     for (i, pixel) in enc_map.iter().map(
@@ -47,7 +47,7 @@ enum DiffuseDirection {
     Forward,
     Backward,
 }
-fn diffuse(im: &Vec<[u8; 3]>, enc_map: &Vec<f64>, direction: DiffuseDirection) -> Vec<[u8; 3]> {
+fn diffuse(im: &Vec<[u8; 3]>, enc_map: &Vec<f32>, direction: DiffuseDirection) -> Vec<[u8; 3]> {
     let enc_map = enc_map
         .iter()
         .map(|&x| x.to_bits())
@@ -79,7 +79,7 @@ fn diffuse(im: &Vec<[u8; 3]>, enc_map: &Vec<f64>, direction: DiffuseDirection) -
 }
 
 #[allow(dead_code)]
-pub fn encode(im: &Vec<[u8; 3]>, x0: f64) -> Vec<[u8; 3]> {
+pub fn encode(im: &Vec<[u8; 3]>, x0: f32) -> Vec<[u8; 3]> {
     let enc_map = generate_map(LogisticMapOptions {
         x: x0,
         r: R,  
@@ -91,7 +91,7 @@ pub fn encode(im: &Vec<[u8; 3]>, x0: f64) -> Vec<[u8; 3]> {
 }
 
 #[allow(dead_code)]
-pub fn decode(im: &Vec<[u8; 3]>, x0: f64) -> Vec<[u8; 3]> {
+pub fn decode(im: &Vec<[u8; 3]>, x0: f32) -> Vec<[u8; 3]> {
     let enc_map = generate_map(LogisticMapOptions {
         x: x0,
         r: R,
